fix: remove require statements that causes support index file to throw an err
