--- helper.js
@@ -1,7 +1,0 @@
-const dataTestNameToSelector = require('./helper/dataTestNameToSelector')
-const parseSelectorWithDataTest = require('./helper/parseSelectorWithDataTest')
-
-module.exports = {
-    dataTestNameToSelector,
-    parseSelectorWithDataTest,
-}
