+++ support.js
@@ -1,4 +1,24 @@
-const parseSelectorWithDataTest = require('./helper/parseSelectorWithDataTest')
+/**
+ * @param {string} dataTestName
+ * @param {string} [prefix] - Default to "dhis2-uicore"
+ * @returns {string}
+ */
+const dataTestNameToSelector = (dataTestName, prefix) => {
+    const defaultPrefix = Cypress.env('dhis2_datatest_prefix') || ''
+    // Empty string is a valid value, so check for undefined
+    const actualPrefix = typeof prefix === 'undefined' ? defaultPrefix : prefix
+    const dataTestId = actualPrefix
+        ? `${actualPrefix}-${dataTestName}`
+        : dataTestName
+
+    return `[data-test="${dataTestId}"]`
+}
+
+const parseSelectorWithDataTest = (selector, prefix) => {
+    return selector.replace(/\{([^}]*)\}/g, (match, dataTestName) =>
+        dataTestNameToSelector(dataTestName, prefix)
+    )
+}
 
 /**
  * Transforms values in curly braces to a data-test selector
